% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allEdges}
\alias{allEdges}
\title{Protein peptide mappings in character vector}
\format{
A character vector:
\describe{
a peptide sequence, then a protein accession and so on
}
}
\source{
https://github.com/OpenMS/OpenMS/blob/develop/share/OpenMS/examples/BSA/BSA1_OMSSA.idXML
}
\usage{
allEdges
}
\description{
A set of peptide obtained mass spectrometry, mapped to protein from uniprot
should be even
}
\examples{
\dontrun{
 allEdges
}
}
\references{
Röst, H. L., Sachsenberg, T., Aiche, S., Bielow, C., Weisser, H., Aicheler, F., Andreotti, S., Ehrlich, H. C., Gutenbrunner, P., Kenar, E., Liang, X., Nahnsen, S., Nilse, L., Pfeuffer, J., Rosenberger, G., Rurik, M., Schmitt, U., Veit, J., Walzer, M., Wojnar, D., … Kohlbacher, O. (2016). OpenMS: a flexible open-source software platform for mass spectrometry data analysis. Nature methods, 13(9), 741–748. https://doi.org/10.1038/nmeth.3959
}
\keyword{datasets}
