% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BSA1_OMSSA.idXML}
\alias{BSA1_OMSSA.idXML}
\title{Protein peptide mappings in idXML file format}
\format{
idXML (see https://abibuilder.informatik.uni-tuebingen.de/archive/openms/Documentation/nightly/html/classOpenMS_1_1IdXMLFile.html)
}
\source{
https://github.com/OpenMS/OpenMS/blob/develop/share/OpenMS/examples/BSA/BSA1_OMSSA.idXML
}
\description{
A set of peptide obtained mass spectrometry, mapped to protein from uniprot
}
