% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayComponent.R
\name{displayComponent}
\alias{displayComponent}
\title{Select an Component to Display}
\usage{
displayComponent(proteinPeptideGraph, displayingComponent)
}
\arguments{
\item{proteinPeptideGraph}{the return from \code{generateBipartiteGraph},
a graph object, containing all peptides and protein in the idXML given}

\item{displayingComponent}{an integer representing the index of the component
that the user want to display}
}
\value{
it returns the component that is displayed
}
\description{
\code{displayComponent} given an index, display that component in the list of
component
}
\details{
It take the return of \code{generateBipartiteGraph} and an integer, and
decompose the graph and uses indexing to select the graph object to display.
What is display is a weakly connected component of the overall graph, the
biological meaning is that only these vertices are relevant to the inference
of any protein in this component.
}
\examples{
wholeGraph <- generateBipartiteGraph(allEdges)
displayComponent(wholeGraph, 1)

}
\references{
Csardi G, Nepusz T (2006). The igraph software package for complex network
research. \emph{InterJournal, Complex Systems,} 1695. https://igraph.org.
}
