% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBipartiteGraph.R
\name{findNumEdges}
\alias{findNumEdges}
\title{finds the number of edges in an idXML file}
\usage{
findNumEdges(peptideIdentificationObjectVector)
}
\arguments{
\item{peptideIdentificationObjectVector}{the vector holding one or more
peptide Identification Objects}
}
\value{
the number of edges
}
\description{
each mapping of protein to peptide counts as one edges, this function counts
that
}
\details{
it looks at the length of each peptideEvidenceVector which is inside of each
peptidehit, which in turn is inside of each peptide identification
}
