% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBipartiteGraph.R
\name{generateBipartiteGraph}
\alias{generateBipartiteGraph}
\title{makes a graph based on peptide and protein}
\usage{
generateBipartiteGraph(preInferenceFilePath, postInferenceFilePath)
}
\arguments{
\item{preInferenceFilePath}{The file path pointing toward the file before
protein inference it contains all possible protein, and all identified
peptides}

\item{postInferenceFilePath}{The file path pointing toward the file after
protein inference it contains all identified proteins}
}
\value{
It will return a graph that include all mapping of protein to peptides
}
\description{
\code{generateBipartiteGraph} Based on what protein mapped to peptide
identified from mass spectrometry, a graph, specifically a bipartite graph, is
drawn.
}
\details{
It first uses the pyopenms package to parse the given file in idXML format,
and iterate through the file to obtain a mapping of protein accession to
peptide sequence, which is then used by function from the igraph package to
generate the graph, then decompose into weakly connected components
Add something about assumptions
}
