% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretproteinidentification-package.R
\docType{package}
\name{interpretproteinidentification-package}
\alias{interpretproteinidentification}
\alias{interpretproteinidentification-package}
\title{interpretproteinidentification: Visually Interprets Protein Identification}
\description{
This package visually interprets protein identification that is based on mass spectrometry peptide identification by connected them to assocaited biological processes and diseases.
}
\keyword{internal}
