% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runinterpretproteinidentification.R
\name{runinterpretproteinidentification}
\alias{runinterpretproteinidentification}
\title{Launch Shiny App for interpretproteinidentification}
\usage{
runinterpretproteinidentification()
}
\value{
No return value but it open up a Shiny page.
}
\description{
A function that launches the Shiny app for interpretproteinidentification
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
interpretproteinidentification::runinterpretproteinidentification()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
