% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectProtein.R
\name{selectProtein}
\alias{selectProtein}
\title{Display a Protein/Peptide and Its Neighbors}
\usage{
selectProtein(graphToSearchIn, vertexIdentifier)
}
\arguments{
\item{graphToSearchIn}{is the component that the user has chosen to displayed
in \code{displayComponent}, selectProtein searches within this component}

\item{vertexIdentifier}{is an identifier for the vertex in the graph, so what
this refers to here would be either a protein accession or peptide sequence}
}
\value{
The sub-graph that include a vertex and all its neighbors
}
\description{
select a protein and all the peptide that it maps to or a peptide and all the
proteins that it maps to, to display. This function expect the output of
\code{generateBipartiteGraph} and \code{displayComponent} to be its input.
}
\details{
Using igraph function, It first finds the vertex object that identifier
refers to, then find the neighbors of that vertex, and finally induce a
sub-graph that include them. When this subgraph is plotted, if the
vertexIdentifier is a protein accession, the gene ontology IDs will be display
at the top of the plot, if it is involve in disease then it will be displayed
at the bottom. (if it is not, then nothing would be displayed).
Also to display all the neighbour of a protein/peptide is to plot what
peptide/protein can be mapped to the vertex in question from
vertexIdentifier
}
\examples{
wholeGraph <- generateBipartiteGraph(allEdges)
oneComponent <- displayComponent(wholeGraph, 1)
selectProtein(oneComponent, "P02769")
selectProtein(wholeGraph, "P02769")
selectProtein(oneComponent, "DDSPDLPK")
selectProtein(wholeGraph, "DDSPDLPK")

}
\references{
Csardi G, Nepusz T (2006). The igraph software package for complex network
research. \emph{InterJournal, Complex Systems,} 1695. https://igraph.org.
}
