% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allEdges}
\alias{allEdges}
\title{Protein peptide mappings in character vector}
\format{
A character vector:
\describe{
peptide sequence, then protein accession and so on
}
}
\source{
https://github.com/OpenMS/OpenMS/blob/develop/share/OpenMS/examples/BSA/BSA1_OMSSA.idXML
}
\usage{
allEdges
}
\description{
A set of peptide obtained mass spectrometry, mapped to protein from uniprot
should be even
}
\examples{
\dontrun{
 allEdges
}
}
\keyword{datasets}
