% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BSA1_OMSSA.idXML}
\alias{BSA1_OMSSA.idXML}
\title{Protein Peptide Mappings in IdXML File Format}
\format{
idXML (see https://abibuilder.informatik.uni-tuebingen.de/archive/
openms/Documentation/nightly/html/classOpenMS_1_1IdXMLFile.html)
}
\source{
https://github.com/OpenMS/OpenMS/blob/develop/share/OpenMS/examples/BSA/BSA1_OMSSA.idXML
}
\description{
A set of peptides obtained mass spectrometry, mapped to proteins from uniprot.
}
\references{
Röst, H. L., Sachsenberg, T., Aiche, S., Bielow, C., Weisser, H.,
Aicheler, F., Andreotti, S., Ehrlich, H. C., Gutenbrunner, P., Kenar, E.,
Liang, X., Nahnsen, S., Nilse, L., Pfeuffer, J., Rosenberger, G.,
Rurik, M., Schmitt, U., Veit, J., Walzer, M., Wojnar, D., …
Kohlbacher, O. (2016). OpenMS: a flexible open-source software
platform for mass spectrometry data analysis. Nature methods, 13(9),
741–748. https://doi.org/10.1038/nmeth.3959
}
