% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayComponent.R
\name{displayComponent}
\alias{displayComponent}
\title{select an component to display}
\usage{
displayComponent(proteinPeptideGraphComponents, displayingComponent)
}
\arguments{
\item{proteinPeptideGraphComponents}{the return from \code{generateBipartiteGraph}}

\item{displayingComponent}{an integer representing the index of the component
that the user want to display}
}
\value{
it returns the component that is displayed
}
\description{
\code{displayComponent} given an index, display that component in the list of
component
}
\details{
It take the return of \code{generateBipartiteGraph} and an integer, and uses
indexing to select the graph object to display
}
