% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBipartiteGraph.R
\name{findNumEdges}
\alias{findNumEdges}
\title{Finds the Number of Edges in an idXML File}
\usage{
findNumEdges(peptideIdentificationObjectVector)
}
\arguments{
\item{peptideIdentificationObjectVector}{the vector holding one or more
peptide Identification Objects}
}
\value{
the number of edges
}
\description{
each mapping of protein to peptide counts as one edges, this function counts
that
}
\details{
it looks at the length of each peptideEvidenceVector which is inside of each
peptidehit, which in turn is inside of each peptide identification
}
\references{
Röst, H. L., Sachsenberg, T., Aiche, S., Bielow, C., Weisser, H.,
Aicheler, F., Andreotti, S., Ehrlich, H. C., Gutenbrunner, P., Kenar,
E., Liang, X., Nahnsen, S., Nilse, L., Pfeuffer, J., Rosenberger, G.,
Rurik, M., Schmitt, U., Veit, J., Walzer, M., Wojnar, D., … Kohlbacher, O.
(2016). OpenMS: a flexible open-source software platform for mass
spectrometry data analysis. \emph{Nature methods, 13}(9), 741–748.
https://doi.org/10.1038/nmeth.3959
}
