% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBipartiteGraph.R
\name{generateBipartiteGraph}
\alias{generateBipartiteGraph}
\title{Generate a Graph based on Peptide and Protein}
\usage{
generateBipartiteGraph(peptideProteinEdgeVector, inferredProteinVector)
}
\arguments{
\item{peptideProteinEdgeVector}{A mapping of protein accession
to peptide sequence in a character vector, either the odd elements are all
proteins and the even elements are all peptide or the odd elements are all
peptides and the even elements are all proteins. There should be an even
number of elements}

\item{inferredProteinVector}{A character vector of all protein that are
inferred (that is considered presented in the pre mass spectrometry sample)
out of all possible proteins}
}
\value{
It will return a igraph object that include all mapping of protein
to peptides
}
\description{
Based on the protein that can be mapped to peptide which are identified from
mass spectrometry, a bipartite graph, is drawn.
}
\details{
It takes a mapping of protein accession to peptide sequence,
which is then used by function from the igraph package to
generate the graph, then decompose into weakly connected components.
This un-directed bipartite graph consist of a set of vertices that represent
proteins and a set of vertices that represent peptides. The edges between
protein vertices and peptide vertices represent protein-peptide matches.
Protein accession, peptide identifiers will be displayed at this layer.
}
\examples{
generateBipartiteGraph(allEdges)

}
\references{
Csardi G, Nepusz T (2006). The igraph software package for complex network
research. \emph{InterJournal, Complex Systems,} 1695. https://igraph.org.
}
