% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretproteinidentification-package.R
\docType{package}
\name{interpretProteinIdentitification}
\alias{interpretProteinIdentitification}
\title{interpretProteinIdentitification: A package for visualizing protein inference results}
\description{
The interpretProteinIdentitification package provides three functions:
generateBipartiteGraph, displayComponent, selectProtein
}
\section{generateBipartiteGraph function}{

The generateBipartiteGraph function take in idXML files and make a graph
}

\section{displayComponent function}{

The displayComponent function take in graph object generated by generateBipartiteGraph and plot one of it components
}

\section{selectProtein function}{

The selectProtein function take in any graph object (either the whole graph from generateBipartiteGraph or a component
from displayComponent) and display only the portion of the graph/component that is connect to the inputted protein
}

