% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBipartiteGraph.R
\name{loadFileIntoVector}
\alias{loadFileIntoVector}
\title{Parses idXML (Does not work due to reticulate not working)}
\usage{
loadFileIntoVector(preInferenceFilePath, postInferenceFilePath, ropenms)
}
\arguments{
\item{preInferenceFilePath}{The file path pointing toward an idXML file
before protein inference, it contains all possible protein, and all identified
peptides}

\item{postInferenceFilePath}{The file path pointing toward an idXML file
after protein inference, it contains all identified/inferred proteins}

\item{ropenms}{thing}
}
\value{
a character vector where every two elements represent an edge
in the graph
}
\description{
using the python package, pyopenms, load the idXML files into an R
object then parse it to obtain what the graph needs
}
\details{
it first take convert R list to python list, load the data into the python
list, then convert the R list back, and iterate through the file to obtain a
mapping of protein to peptides.
An idXML consist of a peptide identification object and a protein
identification object
}
\examples{
\dontrun{
reticulate::conda_create("r-reticulate")
reticulate::conda_install("r-reticulate", "pyopenms") 
ropenms <- DIAlignR::get_ropenms(condaEnv = "r-reticulate", useConda=TRUE)
}
}
