% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBipartiteGraph.R
\name{loadFileIntoVector}
\alias{loadFileIntoVector}
\title{Parses idXML}
\usage{
loadFileIntoVector(idXMLFilePath)
}
\arguments{
\item{idXMLFilePath}{a filepath that point to the idXML file to be parsed}
}
\value{
a list consisting of 2 vectors, one that contains the protein
identification and another one that contains the peptide identification
}
\description{
using the python package, pyopenms, load the idXML files into an R
object
}
\details{
it first take convert R list to python list, load the data into the python
list, then convert the R list back,
}
