% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBipartiteGraph.R
\name{readSQLiteFile}
\alias{readSQLiteFile}
\title{Reads and Parses an SQLite database file}
\usage{
readSQLiteFile(variables)
}
\description{
using RSQLite, it read an OSW file, which is database file that can be read
using with SQLite, this will return the mapping of protein to peptides
}
