% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectProtein.R
\name{selectProtein}
\alias{selectProtein}
\title{display a protein/peptide and its neighbours}
\usage{
selectProtein(graphToSearchIn, vertexIdentifier)
}
\arguments{
\item{graphToSearchIn}{is the component that the user has chosen to displayed
in \code{displayComponent}, selectProtein searches within this component}

\item{vertexIdentifier}{is an identifier for the vertex in the graph, so what
this refers to here would be either a protein accession or peptide sequence}
}
\value{
The sub-graph that include a vertex and all its neighbours
}
\description{
select a protein and all the peptide that it maps to or a peptide and all the
proteins that it maps to, to display
}
\details{
Using igraph function, It first finds the vertex object that identifier
refers to, then find the neighbours of that vertex, and finally induce a
sub-graph that include them. Feature yet to add: displaying protein sequence,
color code peptide and protein sequence, add gene ontology labels.
}
\examples{
componentWanted <- displayComponent(componentList, displayingComponent = 1)
proteinSubgraph <- selectProtein(componentWanted, "P02769|ALBU_BOVIN")
proteinSubgraph <- selectProtein(componentWanted, "DDSPDLPK")
}
