% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{test_data.osw}
\alias{test_data.osw}
\title{Protein Peptide Mappings in OSW File Format}
\format{
transition PQP (see https://abibuilder.informatik.uni-tuebingen.de/
archive/openms/Documentation/release/2.6.0/html/
classOpenMS_1_1TransitionPQPFile.html) for file format description
}
\source{
https://github.com/PyProphet/pyprophet/blob/master/tests/data/test_data.osw
}
\description{
Protein Peptide Mappings in OSW File Format
}
\references{
Teleman, J., Röst, H. L., Rosenberger, G., Schmitt, U.,
Malmström, L., Malmström, J., & Levander, F. (2015).
DIANA--algorithmic improvements for analysis of data-independent acquisition
MS data. Bioinformatics (Oxford, England), 31(4), 555–562.
https://doi.org/10.1093/bioinformatics/btu686
}
